调整界面
将title移动至标题处

+++ tis-client/src/main/java/com/cas/sim/tis/app/hold/ElecCompHoldHandler.java
@@ -96,9 +96,9 @@ public class ElecCompHoldHandler implements HoldHandler {
 		 * 在x-z平面上，每次移动一个单位
 		 */
 		spatial.setLocalTranslation(//
-				contactPoint.x, //
+				(int)contactPoint.x, //
 				contactPoint.y, //
-				contactPoint.z//
+				(int)contactPoint.z//
 		);
 	}
